% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MaissaLajimi}
\alias{MaissaLajimi}
\title{SVM Forecast Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 96 rows and 1 columns.
}
\usage{
MaissaLajimi
}
\description{
SVM Forecast Data
}
\keyword{datasets}
