% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wnn_compute_scaling}
\alias{wnn_compute_scaling}
\title{Compute Scaling Parameters}
\usage{
wnn_compute_scaling(y, X = NULL, m, temp_lags = 1)
}
\arguments{
\item{y}{Numeric vector. Target series}

\item{X}{Numeric vector or matrix. Exogenous variables (optional)}

\item{m}{Integer. Embedding dimension}

\item{temp_lags}{Integer. Number of temperature lags to use (default=1)}
}
\value{
List with scaling parameters
}
\description{
Calculates mean and standard deviation for standardization
}
\examples{
y <- rnorm(100, 150, 20)
params <- wnn_compute_scaling(y, m = 10)
print(params)
}
